Remove FakeMoney command and update packages
Delete the FakeMoney command as it's not serving any real purpose, rather causing confusion among users. In addition, the EXILED and LiteDB packages are updated to their new versions for better performance and security. The types of events are also updated for conformity with the new EXILED version. Lastly, the ItemType.KeycardNTFOfficer is replaced with ItemType.KeycardMTFPrivate as it fits better with the functionality.